refactor: decrease cognitive complexity on vault and consul and place factories in separate file to increase readability
